+++ examples/example/react/src/examples/layout/02/App.tsx
@@ -7,7 +7,7 @@ export default function App() {
   return (
     <Layout className="min-h-64 overflow-hidden rounded border border-gray-300">
       <Header className="!bg-blue-600 !p-4 !text-white">Header</Header>
-      <Content className="!bg-white !p-4">Content</Content>
+      <Content className="!p-4">Content</Content>
       <Footer className="!bg-gray-800 !p-4 !text-white">Footer</Footer>
     </Layout>
   )