+++ examples/example/vue3/src/examples/layout/01/App.vue
@@ -3,7 +3,7 @@ import { Container } from '@expcat/tigercat-vue/Container'
 </script>
 
 <template>
-  <Container max-width="lg" class="rounded border border-gray-200 bg-white p-4">
+  <Container max-width="lg" class="rounded border border-gray-200 p-4">
     响应式居中的 Container 内容
   </Container>
 </template>